+++ rosariosis.sql
@@ -1154,6 +1154,7 @@ CREATE TABLE gradebook_assignments (
     file text,
     default_points integer,
     submission varchar(1),
+    weight integer,
     created_at timestamp DEFAULT current_timestamp,
     updated_at timestamp
 );