+++ CHANGES.md
@@ -25,6 +25,7 @@ Changes in 10.9
 - Remove $count & $has_count_text variables in ListOutput.fnc.php
 - Add pagination option (defaults to false) in ListOutput.fnc.php
 - Add pagination for list > 1000 results in AccessLog.php & ReportCardGrades.php
+- SQL gradebook_assignments table: Add WEIGHT column in Update.fnc.php, rosariosis.sql & rosariosis_mysql.sql
 
 Changes in 10.8.5
 -----------------